+++ src/main/java/com/sheikplastic/repository/GrupoUsuarioRepository.java
@@ -3,5 +3,5 @@ package com.sheikplastic.repository;
 import com.sheikplastic.model.GrupoUsuario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GrupoUsuarioRepository extends JpaRepository<GrupoUsuario, Long> {
+public interface GrupoUsuarioRepository extends JpaRepository<GrupoUsuario, Integer> {
 }
